feat(#2948): extract _meta following nvim pattern
