+++ lua/nvim-tree/renderer/decorator/opened.lua
@@ -27,7 +27,7 @@ end
 
 ---Opened highlight: renderer.highlight_opened_files and node has an open buffer
 ---@param node Node
----@return string|nil group
+---@return string? highlight_group
 function DecoratorOpened:highlight_group(node)
   if self.highlight_range ~= "none" and buffers.is_opened(node) then
     return "NvimTreeOpenedHL"
